Update tests
Update tests according to new `mdance.data` functionality. Now,
directory path is not needed. pytest should work from any path that
accesses the `tests` folder.
